make modal creation to a view instead of modal, export creation theme to it's own theme
